+++ sources/gookto_io/IntersectionNode.cpp
@@ -13,7 +13,7 @@
 #include <iostream>
 
 
-IntersectionNode::IntersectionNode(int type, int id, int num_connection, int num_port) : 
+IntersectionNode::IntersectionNode(int type, long int id, int num_connection, int num_port) : 
 type(type), id(id), numConnections(num_connection), numLinks(num_port)
 {
     //initialize size of connectedLinks --> numLinks
